+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.4
 
  - Effects now track an Entity relative to the given starting Location
+ - Added some simpler versions of EffectManager.start()
 
 ## 3.0